requirements.txt
flask
cohere
scikit-learn
gunicorn
python-dotenv